Added a delete event functionality and a button on the event page (for when the user has logged in) to delete an event.
Added method signature to EventService.java
Added method implementation in EventServiceImp.java
Added delete button next to events shown index.html
Added the event controlers for the button in EventsController.java

This commit adds functionality to delete an event out of the event
database, if the event exists. It also makes it possible for a user that
is logged in to delete an event from shown events.

Further suggestions: make it only allow certain users to do certain
actions. For example, only the event creator (or people working on that
event) can delete an event.

Resolve: #12