Allow new events to be added
correct routes implemented in EventController to send user to events/new
for form submission and to POST the form.

new.html is the page user sent to to input the new event information.
This minimum viable product implementation has the user select a venue
from a dropdown, with no option to add a new venue.